fix(core): create separate pseudo pty implementation for mac
