+++ packages/nx/src/native/index.js
@@ -246,19 +246,19 @@ if (!nativeBinding) {
   throw new Error(`Failed to load native binding`)
 }
 
-const { expandOutputs, getFilesForOutputs, remove, copy, RustPseudoTerminal, ChildProcess, hashArray, hashFile, ImportResult, findImports, transferProjectGraph, HashPlanner, TaskHasher, EventType, Watcher, WorkspaceContext, WorkspaceErrors, testOnlyTransferFileMap } = nativeBinding
+const { expandOutputs, getFilesForOutputs, remove, copy, hashArray, hashFile, ImportResult, findImports, transferProjectGraph, ChildProcess, RustPseudoTerminal, HashPlanner, TaskHasher, EventType, Watcher, WorkspaceContext, WorkspaceErrors, testOnlyTransferFileMap } = nativeBinding
 
 module.exports.expandOutputs = expandOutputs
 module.exports.getFilesForOutputs = getFilesForOutputs
 module.exports.remove = remove
 module.exports.copy = copy
-module.exports.RustPseudoTerminal = RustPseudoTerminal
-module.exports.ChildProcess = ChildProcess
 module.exports.hashArray = hashArray
 module.exports.hashFile = hashFile
 module.exports.ImportResult = ImportResult
 module.exports.findImports = findImports
 module.exports.transferProjectGraph = transferProjectGraph
+module.exports.ChildProcess = ChildProcess
+module.exports.RustPseudoTerminal = RustPseudoTerminal
 module.exports.HashPlanner = HashPlanner
 module.exports.TaskHasher = TaskHasher
 module.exports.EventType = EventType
