+++ packages/nx/src/native/pseudo_terminal/command/unix.rs
