+++ packages/nx/src/native/pseudo_terminal/command/windows.rs
