+++ packages/nx/src/native/pseudo_terminal/mod.rs
@@ -1,0 +1,11 @@
+#[cfg_attr(windows, path = "command/windows.rs")]
+#[cfg_attr(not(windows), path = "command/unix.rs")]
+mod os;
+
+mod pseudo_terminal;
+
+pub mod child_process;
+
+#[cfg_attr(target_os = "macos", path = "mac.rs")]
+#[cfg_attr(not(target_os = "macos"), path = "non-mac.rs")]
+pub mod rust_pseudo_terminal;
